% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayz_check_modelterms.R
\name{bayz_check_modelterms}
\alias{bayz_check_modelterms}
\title{Bayz check model terms}
\usage{
bayz_check_modelterms(model)
}
\arguments{
\item{model}{A formula describing the model.}
}
\value{
A fitted bayz model
}
\description{
Checks if a model is prepared to run through bayz.
Detects if the support functions, fixf(), ranf() ... etc is being used.
Lists all of the
}
