% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.bayz}
\alias{coef.bayz}
\title{Extract posterior means and SDs for all or a selection of coefficients}
\usage{
\method{coef}{bayz}(object, which = NULL, splitLabels = TRUE, ...)
}
\arguments{
\item{object}{A bayz model output}

\item{which}{A vector TRUE/FALSE to indicate which model coefficients to extract (should match the rows
in the $Parameters table in a bayz output object). It can be omitted to extract all coefficients.}

\item{splitLabels}{Whether labels that contain % (in interactions) such as a%b should be split
in multiple columns (default TRUE).}

\item{...}{Additional parameters.}
}
\value{
a list with one member (a data frame) for each coefficient
}
\description{
Extract from the Bayz output all or a selection of coefficients (posterior mean and SD), i.e. estimates from the
fx(), rg() and rn() terms.
}
