% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef.R
\name{fixef.bayz}
\alias{fixef.bayz}
\title{Extract posterior means and SDs for `fixed´ effects}
\usage{
\method{fixef}{bayz}(object, splitLabels = TRUE, ...)
}
\arguments{
\item{object}{A bayz model output}

\item{splitLabels}{Whether labels that contain % (in interactions) such as a%b should be split
in multiple columns (default TRUE).}

\item{...}{Additional parameters passed onto the Model function.}
}
\value{
a list with one member (a data frame) for each fixed effect
}
\description{
Extract from the Bayz output the coefficients (posterior mean and SD) for terms fitted with fx() and rg().
With the default splitLabels=TRUE, labels for interactions such as a%b will be split in multiple columns.
}
