% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixf.R
\name{fixf}
\alias{fixf}
\title{Fixed variable}
\usage{
fixf(x)
}
\arguments{
\item{x}{Variable vector which should be fixed in a model.}
}
\value{
Factorized version of the vector
}
\description{
Sets a variable to a fixed variable in a linear mixed model.
}
