% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freg.R
\name{freg}
\alias{freg}
\title{Fixed-effect regression on continuous covariable}
\usage{
freg(x)
}
\arguments{
\item{x}{Vector of covariates (will be forced to be numeric) to fit in the model as a fixed regression.}
}
\value{
Numeric version of the vector
}
\description{
Fixed-effect regression on continuous covariable
}
