% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichi.R
\name{ichi}
\alias{ichi}
\title{prior function}
\usage{
ichi(scale = 0, df = -2)
}
\arguments{
\item{scale}{Scale}

\item{df}{Degrees of freedom}
}
\value{
Factorized version of the vector
}
\description{
Sets a variable to a fixed variable in a linear mixed model.
}
