% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor.R
\name{mfactor}
\alias{mfactor}
\title{mfactor function to build multi-column/matrix factor from list of single factors}
\usage{
mfactor(...)
}
\arguments{
\item{...}{list of factors to be merge into one multi-column factor}
}
\value{
object of type mfactor
}
\description{
mfactor function to build multi-column/matrix factor from list of single factors
}
