% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Bayesian mixed model}
\usage{
mm(data = NULL, chain = NULL, ...)
}
\arguments{
\item{data}{Data frame to collect data from}

\item{...}{Additional parameters passed onto the Model function (for example )}
}
\value{
Fitted mixed model
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}
