% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmg.R
\name{mmg}
\alias{mmg}
\title{Bayesian mixed model with genotypes}
\usage{
mmg(data = NULL, geno = NULL, ...)
}
\arguments{
\item{data}{Data frame to collect data from}

\item{geno}{Table with genotypes}

\item{...}{Additional parameters passed onto the Model function (for example )}
}
\value{
Fitted mixed model
}
\description{
Bayesian mixed model with table of genotypes for random regression (rrBLUP/SNP-BLUP) fit
}
