% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bayz}
\alias{plot.bayz}
\title{Plotting a bayz object}
\usage{
\method{plot}{bayz}(x, ...)
}
\arguments{
\item{x}{bayz object}

\item{...}{additional parameters}
}
\value{
nothing, plots the convergence of the parameters
}
\description{
Produced summary statistics of the bayz object and returns a summarybayz object.
}
