% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bayz}
\alias{print.bayz}
\title{Print function for bayz output class}
\usage{
\method{print}{bayz}(x, ...)
}
\arguments{
\item{x}{bayz object}

\item{...}{additional parameters}
}
\value{
Nothing, only prints
}
\description{
This method is usually not seen or used directly by the user, but allows to see a compact version of the output
in the console without overflowing the screen, for instance when running bayz() without catching the output, or
to quickly see error messages without the need to use summary().
}
