% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{print.summarybayz}
\alias{print.summarybayz}
\title{Print function for 'summarybayz' class}
\usage{
\method{print}{summarybayz}(x, ...)
}
\arguments{
\item{x}{bayz object}

\item{...}{additional parameters}
}
\value{
Nothing, only prints
}
\description{
Prints the bayz class
}
