% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.bayz}
\alias{ranef.bayz}
\title{Extract posterior means and SDs for `random´ effects}
\usage{
\method{ranef}{bayz}(object, splitLabels = TRUE, ...)
}
\arguments{
\item{object}{A bayz model output}

\item{splitLabels}{Whether labels that contain % (in interactions) such as a%b should be split
in multiple columns (default TRUE).}

\item{...}{Additional parameters passed onto the Model function.}
}
\value{
a list with one member (a data frame) for each random effect
}
\description{
Extract from the Bayz output the coefficients (posterior mean and SD) for rn() model terms.
}
