% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranf.R
\name{ranf}
\alias{ranf}
\title{Specify fit of a random factor, optionally with a kernel/similarity matrix}
\usage{
ranf(x, V = NULL, rrankpct = 99, prior = NULL)
}
\arguments{
\item{x}{Variable (forced to be factor) to be used as random effect in the model.}

\item{V}{Kernel/similarity/var-covar matrix to be included so that x~N(0,Vsigma^2). Uses eigen decomposition.}

\item{rrankpct}{Reduced Rank threshold for including eigenvectors, given in percent (default 99)}

\item{prior}{Prior information to be added.}
}
\value{
Object storing the factor information.
}
\description{
Specify fit of a random factor, optionally with a kernel/similarity matrix
}
