% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbayz.R
\name{rbayz}
\alias{rbayz}
\title{Bayz function}
\usage{
rbayz(model, data = NULL, chain = NULL)
}
\arguments{
\item{model}{A formula describing the model.}

\item{data}{Data frame to collect data from}

\item{fct}{Model function}

\item{...}{Additional parameters passed onto the Model function (for example )}
}
\value{
A fitted bayz model
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}
