% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bayz}
\alias{summary.bayz}
\title{Summary of traced parameters from bayz model fit}
\usage{
\method{summary}{bayz}(object, maxLevel = 10, HPDprob = 0.95, ...)
}
\arguments{
\item{object}{bayz output object}

\item{HPDprob}{probability for the Highest Posterior Density intervals (default 0.95)}

\item{...}{additional parameters}
}
\value{
summarybayz object
}
\description{
Posterior summary statistics including HPD interval and convergence diagnostics of traced parameters
    from a bayz model fit, returns a summarybayz object.
}
