% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayz.R
\name{bayz}
\alias{bayz}
\title{Bayz function}
\usage{
bayz(model, data = NULL, fct = NULL, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{data}{Data frame to collect data from}

\item{fct}{Function to base the model on}

\item{...}{Additional data, for example a data frame.}
}
\value{
A fitted bayz model
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}
