% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.R
\name{ranef.bayz}
\alias{ranef.bayz}
\title{Extract posterior means and SDs for `random´ effects}
\usage{
ranef.bayz(object, splitLabels = TRUE, ...)
}
\arguments{
\item{object}{A bayz model output}

\item{splitLabels}{Whether labels that contain % (in interactions) such as a%b should be split
in multiple columns (default TRUE).}

\item{...}{Additional parameters passed onto the Model function.}
}
\value{

}
\description{
Extract from the Bayz output the estimates (posterior mean and SD) for those terms that would be
called random in mixed models,
i.e. those that have shrinkage priors such as a Normal prior, but also other shrinkage priors.
}
