% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranf.R
\name{ranf}
\alias{ranf}
\title{Set Random variable}
\usage{
ranf(x, cor = NULL, prior = NULL)
}
\arguments{
\item{x}{Variable vector which should be set to random in a model.}

\item{cor}{Correlation matrix to be included. Default uses eigen decomposition}

\item{prior}{Prior information to be added.}
}
\value{
Random variable information.
}
\description{
Sets a variable to a random variable in a linear mixed model.
}
