% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbayz.R
\name{rbayz}
\alias{rbayz}
\title{Bayz test function}
\usage{
rbayz(model, data = NULL, chain = NULL)
}
\arguments{
\item{model}{A formula describing the model.}

\item{data}{Data frame to collect data from}

\item{chain}{Number of iteration/sample to generate.}
}
\value{
A fitted bayz model
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}
