% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbayz_cpp}
\alias{rbayz_cpp}
\title{Core bayz C++ function call}
\usage{
rbayz_cpp(modelFrame, chain)
}
\arguments{
\item{chain}{Three value vector containing the number of iterations c(1000,100,10)}

\item{modelframe}{Generated}

\item{fct}{Model function}

\item{...}{Additional parameters passed on}
}
\value{
A fitted bayz model
}
\description{
Fits a model based on the given modelframe, using Markovian Coalescens
}
