% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bayz}
\alias{summary.bayz}
\title{Summary of bayz model fit}
\usage{
\method{summary}{bayz}(object, maxLevel = 10, HPDprob = 0.95, ...)
}
\arguments{
\item{object}{bayz output object}

\item{maxLevel}{controls the inclusion of estimates in the "Main estimates" table by selecting
only the model-terms with less <= maxlevel coefficients}

\item{HPDprob}{the probability for the Highest Posterior Density intervals reported in the
table "Convergence diagnostics"}

\item{...}{additional parameters}
}
\value{
summarybayz object
}
\description{
Produced summary statistics of the bayz object and returns a summarybayz object.
}
