% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bayz}
\alias{summary.bayz}
\title{Print function for bayz class}
\usage{
\method{summary}{bayz}(x, ...)
}
\arguments{
\item{x}{bayz object}

\item{...}{additional parameters}
}
\value{
Nothing, only prints
}
\description{
Prints the bayz class
}
